examples/chalice-mail-test1/requirements.txt
chalice-mail==0.0.1